Validador de CPF

+++ app/src/main/java/br/edu/ifpr/bsi/prefeiturainterativa/helpers/CpfValidatorHelper.java
@@ -18,7 +18,7 @@ public class CpfValidatorHelper {
         sm = 0;
         peso = 10;
         for (i = 0; i < 9; i++) {
-            num = (int) (CPF.charAt(i) - 48);
+            num = CPF.charAt(i) - 48;
             sm = sm + (num * peso);
             peso = peso - 1;
         }
@@ -31,7 +31,7 @@ public class CpfValidatorHelper {
         sm = 0;
         peso = 11;
         for (i = 0; i < 10; i++) {
-            num = (int) (CPF.charAt(i) - 48);
+            num = CPF.charAt(i) - 48;
             sm = sm + (num * peso);
             peso = peso - 1;
         }
